fixes layout bug when leaving maximized window state #7
Changing a maximized window to normal state fires 'move' and 'unmaximize' (not used) events. In that moment a redraw and reflow of the web document is forced so that the responsive layout is recalculated and applied properly. This is done via hiding and - after a delay - showing the root element. The move event is also continuously fired when dragging the window around. Since this event fires so often, constantly redrawing the page with 'show/hide' results in distracting flickering. Therefore it is saved in a variable if the window was maximized and not yet rerendered. The rerender is debounced to prevent the flickering. While in the window-still-dragging-state the window is already rendered blurred with the wrong fullscreen layout to give the user a feedback that a rerender has not yet happened and he has to release the drag to see the content